add regression test
