+++ regression-test/data/nereids_rules_p0/mv/agg_with_roll_up/aggregate_with_roll_up.out
@@ -1,0 +1,123 @@
+-- This file is automatically generated. You should know what you did if you want to edit this
+-- !query13_0_before --
+3	3	2023-12-11	43.20	43.20	43.20	1	0
+
+-- !query13_0_after --
+3	3	2023-12-11	43.20	43.20	43.20	1	0
+
+-- !query14_0_before --
+2	3	2023-12-08	20.00	10.50	9.50	2	0
+2	3	2023-12-12	\N	\N	\N	1	0
+2	4	2023-12-10	\N	\N	\N	1	0
+3	3	2023-12-11	\N	\N	\N	1	0
+4	3	2023-12-09	\N	\N	\N	1	0
+
+-- !query14_0_after --
+2	3	2023-12-08	20.00	10.50	9.50	2	0
+2	3	2023-12-12	\N	\N	\N	1	0
+2	4	2023-12-10	\N	\N	\N	1	0
+3	3	2023-12-11	\N	\N	\N	1	0
+4	3	2023-12-09	\N	\N	\N	1	0
+
+-- !query15_0_before --
+3	3	2023-12-11	43.20	43.20	43.20	1	0
+
+-- !query15_0_after --
+3	3	2023-12-11	43.20	43.20	43.20	1	0
+
+-- !query16_0_before --
+3	3	2023-12-11	43.20	43.20	43.20	1	0
+
+-- !query16_0_after --
+3	3	2023-12-11	43.20	43.20	43.20	1	0
+
+-- !query17_0_before --
+3	3	2023-12-11	43.20	43.20	43.20	1	0
+
+-- !query17_0_after --
+3	3	2023-12-11	43.20	43.20	43.20	1	0
+
+-- !query18_0_before --
+
+-- !query18_0_after --
+
+-- !query19_0_before --
+2	3	2023-12-08	20.00
+2	3	2023-12-12	57.40
+2	4	2023-12-10	46.00
+
+-- !query19_0_after --
+2	3	2023-12-08	20.00
+2	3	2023-12-12	57.40
+2	4	2023-12-10	46.00
+
+-- !query20_0_before --
+2023-12-08	3	2023-12-08	20.00	10.50	9.50	2	0
+2023-12-09	3	2023-12-09	11.50	11.50	11.50	1	0
+2023-12-10	4	2023-12-10	46.00	33.50	12.50	2	0
+2023-12-11	3	2023-12-11	43.20	43.20	43.20	1	0
+2023-12-12	3	2023-12-12	57.40	56.20	1.20	2	0
+
+-- !query20_0_after --
+2023-12-08	3	2023-12-08	20.00	10.50	9.50	2	0
+2023-12-09	3	2023-12-09	11.50	11.50	11.50	1	0
+2023-12-10	4	2023-12-10	46.00	33.50	12.50	2	0
+2023-12-11	3	2023-12-11	43.20	43.20	43.20	1	0
+2023-12-12	3	2023-12-12	57.40	56.20	1.20	2	0
+
+-- !query21_0_before --
+2	3	2023-12-08	20.00	10.50	9.50	2	0
+2	3	2023-12-12	57.40	56.20	1.20	2	0
+2	4	2023-12-10	46.00	33.50	12.50	2	0
+
+-- !query21_0_after --
+2	3	2023-12-08	20.00	10.50	9.50	2	0
+2	3	2023-12-12	57.40	56.20	1.20	2	0
+2	4	2023-12-10	46.00	33.50	12.50	2	0
+
+-- !query22_0_before --
+2	3	2023-12-08	20.00	10.50	9.50	2	0
+2	3	2023-12-12	57.40	56.20	1.20	2	0
+2	4	2023-12-10	46.00	33.50	12.50	2	0
+3	3	2023-12-11	43.20	43.20	43.20	1	0
+4	3	2023-12-09	11.50	11.50	11.50	1	0
+
+-- !query22_0_after --
+2	3	2023-12-08	20.00	10.50	9.50	2	0
+2	3	2023-12-12	57.40	56.20	1.20	2	0
+2	4	2023-12-10	46.00	33.50	12.50	2	0
+3	3	2023-12-11	43.20	43.20	43.20	1	0
+4	3	2023-12-09	11.50	11.50	11.50	1	0
+
+-- !query23_0_before --
+2	3	2023-12-08	20.00	10.50	9.50	2	0
+
+-- !query23_0_after --
+2	3	2023-12-08	20.00	10.50	9.50	2	0
+
+-- !query24_0_before --
+3	2023-12-08	20.00	10.50	9.50	2	0
+3	2023-12-09	11.50	11.50	11.50	1	0
+3	2023-12-11	43.20	43.20	43.20	1	0
+3	2023-12-12	57.40	56.20	1.20	2	0
+
+-- !query24_0_after --
+3	2023-12-08	20.00	10.50	9.50	2	0
+3	2023-12-09	11.50	11.50	11.50	1	0
+3	2023-12-11	43.20	43.20	43.20	1	0
+3	2023-12-12	57.40	56.20	1.20	2	0
+
+-- !query25_0_before --
+2	3	2023-12-08	20.00	10.50	9.50	2
+2	3	2023-12-12	57.40	56.20	1.20	2
+2	4	2023-12-10	46.00	33.50	12.50	2
+3	3	2023-12-11	43.20	43.20	43.20	1
+4	3	2023-12-09	11.50	11.50	11.50	1
+
+-- !query25_0_after --
+2	3	2023-12-08	20.00	10.50	9.50	2
+2	3	2023-12-12	57.40	56.20	1.20	2
+2	4	2023-12-10	46.00	33.50	12.50	2
+3	3	2023-12-11	43.20	43.20	43.20	1
+4	3	2023-12-09	11.50	11.50	11.50	1
+
