+++ regression-test/data/nereids_rules_p0/mv/agg_without_roll_up/aggregate_without_roll_up.out
@@ -1,0 +1,127 @@
+-- This file is automatically generated. You should know what you did if you want to edit this
+-- !query1_0_before --
+
+-- !query1_0_after --
+1	yy	0	0	77.50	33.50	9.50	5
+2	mi	0	0	57.40	56.20	1.20	2
+2	mm	0	0	43.20	43.20	43.20	1
+
+-- !query1_2_before --
+1	yy	0	0	0	0	0	0	0	0	0	0	0	0	0	0	0
+2	mi	0	0	0	0	0	0	0	0	0	0	0	0	0	0	0
+2	mm	0	0	0	0	0	0	0	0	0	0	0	0	0	0	0
+
+-- !query1_2_after --
+1	yy	0	0	0	0	0	0	0	0	0	0	0	0	0	0	0
+2	mi	0	0	0	0	0	0	0	0	0	0	0	0	0	0	0
+2	mm	0	0	0	0	0	0	0	0	0	0	0	0	0	0	0
+
+-- !query2_0_before --
+1	yy	0	0	0	0	0	0	0	0	0	0	0
+2	mi	0	0	0	0	0	0	0	0	0	0	0
+2	mm	0	0	0	0	0	0	0	0	0	0	0
+
+-- !query2_0_after --
+1	yy	0	0	0	0	0	0	0	0	0	0	0
+2	mi	0	0	0	0	0	0	0	0	0	0	0
+2	mm	0	0	0	0	0	0	0	0	0	0	0
+
+-- !query3_0_before --
+0	0	0	0	0	0	0	0	0	0	0	0	0	0	0
+
+-- !query3_0_after --
+0	0	0	0	0	0	0	0	0	0	0	0	0	0	0
+
+-- !query13_0_before --
+2	3	2023-12-08	20.00	10.50	9.50	2	0
+2	3	2023-12-12	57.40	56.20	1.20	2	0
+2	4	2023-12-10	46.00	33.50	12.50	2	0
+
+-- !query13_0_after --
+2	3	2023-12-08	20.00	10.50	9.50	2	0
+2	3	2023-12-12	57.40	56.20	1.20	2	0
+2	4	2023-12-10	46.00	33.50	12.50	2	0
+
+-- !query14_0_before --
+2	3	\N	\N	\N	\N	1	0
+2	3	2023-12-08	20.00	10.50	9.50	2	0
+2	4	\N	\N	\N	\N	1	0
+3	3	\N	\N	\N	\N	1	0
+4	3	\N	\N	\N	\N	1	0
+
+-- !query14_0_after --
+2	3	\N	\N	\N	\N	1	0
+2	3	2023-12-08	20.00	10.50	9.50	2	0
+2	4	\N	\N	\N	\N	1	0
+3	3	\N	\N	\N	\N	1	0
+4	3	\N	\N	\N	\N	1	0
+
+-- !query15_0_before --
+2	3	2023-12-08	20.00	10.50	9.50	2	0
+2	3	2023-12-12	57.40	56.20	1.20	2	0
+2	4	2023-12-10	46.00	33.50	12.50	2	0
+3	3	2023-12-11	43.20	43.20	43.20	1	0
+
+-- !query15_0_after --
+2	3	2023-12-08	20.00	10.50	9.50	2	0
+2	3	2023-12-12	57.40	56.20	1.20	2	0
+2	4	2023-12-10	46.00	33.50	12.50	2	0
+3	3	2023-12-11	43.20	43.20	43.20	1	0
+4	3	2023-12-09	11.50	11.50	11.50	1	0
+
+-- !query16_0_before --
+2	3	2023-12-08	20.00	10.50	9.50	2	0
+2	3	2023-12-12	57.40	56.20	1.20	2	0
+2	4	2023-12-10	46.00	33.50	12.50	2	0
+3	3	2023-12-11	43.20	43.20	43.20	1	0
+4	3	2023-12-09	11.50	11.50	11.50	1	0
+
+-- !query16_0_after --
+2	3	2023-12-08	20.00	10.50	9.50	2	0
+2	3	2023-12-12	57.40	56.20	1.20	2	0
+2	4	2023-12-10	46.00	33.50	12.50	2	0
+3	3	2023-12-11	43.20	43.20	43.20	1	0
+4	3	2023-12-09	11.50	11.50	11.50	1	0
+
+-- !query17_0_before --
+3	3	2023-12-11	43.20	43.20	43.20	1	0
+
+-- !query17_0_after --
+3	3	2023-12-11	43.20	43.20	43.20	1	0
+
+-- !query18_0_before --
+
+-- !query18_0_after --
+
+-- !query18_1_before --
+
+-- !query18_1_after --
+
+-- !query18_2_before --
+
+-- !query18_2_after --
+
+-- !query19_0_before --
+2	3	2023-12-08	20.00	10.50	9.50	2
+2	3	2023-12-12	57.40	56.20	1.20	2
+2	4	2023-12-10	46.00	33.50	12.50	2
+3	3	2023-12-11	43.20	43.20	43.20	1
+4	3	2023-12-09	11.50	11.50	11.50	1
+
+-- !query19_0_after --
+2	3	2023-12-08	20.00	10.50	9.50	2
+2	3	2023-12-12	57.40	56.20	1.20	2
+2	4	2023-12-10	46.00	33.50	12.50	2
+3	3	2023-12-11	43.20	43.20	43.20	1
+4	3	2023-12-09	11.50	11.50	11.50	1
+
+-- !query19_1_before --
+4	1	77.50
+4	2	43.20
+6	2	57.40
+
+-- !query19_1_after --
+4	1	77.50
+4	2	43.20
+6	2	57.40
+
