+++ regression-test/data/nereids_rules_p0/mv/join/inner/inner_join.out
@@ -1,0 +1,267 @@
+-- This file is automatically generated. You should know what you did if you want to edit this
+-- !query1_0_before --
+4
+4
+4
+4
+4
+4
+6
+6
+
+-- !query1_0_after --
+4
+4
+4
+4
+4
+4
+6
+6
+
+-- !query1_1_before --
+4
+4
+4
+4
+6
+6
+6
+6
+
+-- !query1_1_after --
+4
+4
+4
+4
+6
+6
+6
+6
+
+-- !query1_2_before --
+4
+4
+4
+4
+4
+4
+6
+6
+
+-- !query1_2_after --
+4
+4
+4
+4
+4
+4
+6
+6
+
+-- !query1_3_before --
+1	1
+1	1
+1	1
+1	1
+1	1
+2	2
+2	2
+2	2
+
+-- !query1_3_after --
+1	1
+1	1
+1	1
+1	1
+1	1
+2	2
+2	2
+2	2
+
+-- !query1_4_before --
+4
+4
+4
+4
+6
+6
+6
+6
+
+-- !query1_4_after --
+4
+4
+4
+4
+6
+6
+6
+6
+
+-- !query2_0_before --
+
+-- !query2_0_after --
+
+-- !query2_1_before --
+4
+4
+4
+4
+4
+4
+6
+6
+
+-- !query2_1_after --
+4
+4
+4
+4
+4
+4
+6
+6
+
+-- !query2_3_before --
+4
+4
+4
+4
+6
+6
+6
+6
+
+-- !query2_3_after --
+4
+4
+4
+4
+6
+6
+6
+6
+
+-- !query3_0_before --
+4
+4
+4
+4
+4
+4
+6
+6
+
+-- !query3_0_after --
+4
+4
+4
+4
+4
+4
+6
+6
+
+-- !query3_1_before --
+4
+4
+4
+4
+4
+4
+6
+6
+
+-- !query3_1_after --
+4
+4
+4
+4
+4
+4
+6
+6
+
+-- !query3_2_before --
+4
+4
+4
+4
+4
+4
+6
+6
+
+-- !query3_2_after --
+4
+4
+4
+4
+4
+4
+6
+6
+
+-- !query3_3_before --
+4
+4
+4
+4
+6
+6
+6
+6
+
+-- !query3_3_after --
+4
+4
+4
+4
+6
+6
+6
+6
+
+-- !query4_0_before --
+4
+4
+4
+4
+4
+4
+
+-- !query4_0_after --
+4
+4
+4
+4
+4
+4
+6
+6
+
+-- !query5_0_before --
+4
+4
+4
+4
+4
+4
+6
+6
+
+-- !query5_0_after --
+4
+4
+4
+4
+4
+4
+6
+6
+
+-- !query10_0_before --
+
+-- !query10_0_after --
+
