+++ regression-test/data/nereids_rules_p0/mv/join/left_outer/outer_join.out
@@ -1,0 +1,223 @@
+-- This file is automatically generated. You should know what you did if you want to edit this
+-- !query1_0_before --
+4
+4
+4
+4
+4
+4
+6
+6
+
+-- !query1_0_after --
+4
+4
+4
+4
+4
+4
+6
+6
+
+-- !query1_1_before --
+4
+4
+4
+4
+4
+4
+4
+4
+6
+6
+6
+6
+
+-- !query1_1_after --
+4
+4
+4
+4
+4
+4
+4
+4
+6
+6
+6
+6
+
+-- !query1_2_before --
+4
+4
+4
+4
+4
+4
+6
+6
+
+-- !query1_2_after --
+4
+4
+4
+4
+4
+4
+6
+6
+
+-- !query1_3_before --
+1	1
+1	1
+1	1
+1	1
+1	1
+2	2
+2	2
+2	2
+
+-- !query1_3_after --
+1	1
+1	1
+1	1
+1	1
+1	1
+2	2
+2	2
+2	2
+
+-- !query2_0_before --
+
+-- !query2_0_after --
+
+-- !query2_1_before --
+4
+4
+4
+4
+4
+4
+6
+6
+
+-- !query2_1_after --
+4
+4
+4
+4
+4
+4
+6
+6
+
+-- !query3_0_before --
+4
+4
+4
+4
+4
+4
+6
+6
+
+-- !query3_0_after --
+4
+4
+4
+4
+4
+4
+6
+6
+
+-- !query3_1_before --
+4
+4
+4
+4
+4
+4
+6
+6
+
+-- !query3_1_after --
+4
+4
+4
+4
+4
+4
+6
+6
+
+-- !query3_2_before --
+4
+4
+4
+4
+4
+4
+6
+6
+
+-- !query3_2_after --
+4
+4
+4
+4
+4
+4
+6
+6
+
+-- !query4_0_before --
+4
+4
+
+-- !query4_0_after --
+4
+4
+
+-- !query5_0_before --
+4
+4
+4
+4
+4
+4
+6
+6
+
+-- !query5_0_after --
+4
+4
+4
+4
+4
+4
+6
+6
+
+-- !query6_0_before --
+2	3	2023-12-08
+2	3	2023-12-08
+2	3	2023-12-12
+2	4	2023-12-10
+3	3	2023-12-11
+4	3	2023-12-09
+
+-- !query6_0_after --
+2	3	2023-12-08
+2	3	2023-12-08
+2	3	2023-12-12
+2	4	2023-12-10
+3	3	2023-12-11
+4	3	2023-12-09
+
+-- !query7_0_before --
+3	3	2023-12-11
+
+-- !query7_0_after --
+3	3	2023-12-11
+
