+++ regression-test/suites/nereids_rules_p0/mv/join/inner/inner_join.groovy
