+++ regression-test/suites/nereids_rules_p0/mv/join/left_outer/outer_join.groovy
